+++ packages/mock-block-dock/src/debug-view/dev-tools/json-view.tsx
