+++ packages/mock-block-dock/src/debug-view/dev-tools/tab-panel.tsx
